# perf(tsdb): optimize ExpandPostings with slice pre-allocation